+++ src/components/common/Logo/index.tsx
@@ -5,12 +5,19 @@ import React from "react";
 type LogoProps = {
     typographyProps?: TypographyProps,
     style?: Object
+    onClick?: () => void;
 }
 
 const Logo = (props: LogoProps) => {
     const classes = useStyles();
 
-    return <Typography variant="h2" className={classes.root} style={props.style} {...props.typographyProps}>Circles</Typography>
+    const style = {
+        ...props.style,
+        // Set cursor to pointer if the Logo is clickable.
+        cursor: props.onClick ? "pointer" : "initial"
+    }
+
+    return <Typography onClick={props.onClick} variant="h2" className={classes.root} style={style} {...props.typographyProps}>Circles</Typography>
 }
 
 export default Logo;