+++ src/components/pages/Home/index.tsx
@@ -33,6 +33,10 @@ function Home() {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 
+  const handleNavigation = (path: string) => () =>
+    history.push(path)
+
+
   return (
     <>
       <section className={classes.root}>
@@ -42,7 +46,7 @@ function Home() {
         <div className={classes.container}>
           <section className={classes.topMenu}>
             {/* top Menu */}
-            <Logo style={{ fontSize: "56px", margin: "16px 0" }} />
+            <Logo style={{ fontSize: "56px", margin: "16px 0" }} onClick={handleNavigation('/home/')} />
             <Flex justifyContent="flex-end" alignItems="center" className={classes.menuBar}>
               <SettingsDrawer />
               <Avatar className={classes.avatar} alt="user" src={currentAccount?.avatar_url} onClick={openProfileDrawer} />