+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Logo is now clickable, redirects to the home view.
+
 ### Changed
 - Settings drawer will be kept mounted to avoid expensive refetching of background thumbnails.